売却処理をMarketに移動
